+++ inc/requires/mkvmergeItem.php
@@ -11,7 +11,7 @@ function mkvmergeItem($file, $fileorig, $options, $info)
   if (empty($file) || empty($options) || empty($info)) return (array([], [], []));
   if (file_exists($file['filename'] . $mkvm_ext)) unlink ($file['filename'] . $mkvm_ext); // leftover detected
   if (!isset($fileorig)) $fileorig =[];
-  if (`which mkvmerge 2> /dev/null` && !$options['args']['nomkvmerge'] && !isStopped($options) && !$options['args']['test']) {
+  if (`which mkvmerge 2> /dev/null` && !$options['args']['nomkvmerge'] && !isStopped($options)) {
     if (!$info['format']['mkvmerged'] && !$info['format']['exclude'] && !$options['args']['exclude']) {
       print ansiColor("blue") . "Preprocessing: " . ansiColor("red") . $file['basename'] . ansiColor() . " (" . formatBytes(filesize($file['basename']), 2, true) . ")\n" . ansiColor();
       $cmdln    = "mkvmerge";
@@ -28,6 +28,9 @@ function mkvmergeItem($file, $fileorig, $options, $info)
         $cmdln .=
           " --audio-tracks " . $info['audio']['index'] .
           " --track-order " . "0:" . $info['video']['index'] . "," . "1:" . $info['audio']['index'];
+      } else {
+        $cmdln .=
+        " --audio-tracks " . $options['args']['language'];
       }
       $cmdln .=
         " --no-attachments";
@@ -41,30 +44,30 @@ function mkvmergeItem($file, $fileorig, $options, $info)
       if ($options['args']['verbose']) {
         print "\n\n" . ansiColor("green") . "$cmdln\n" . ansiColor();
       }
-
-      system("$cmdln 2>&1", $status);
-      if ($status == 255) {
-        // status(255) => CTRL-C
-        // Restore and Cleanup
-        if(!empty($fileorig) && !empty($file)) {
-          rename($fileorig['basename'], $file['basename']); 
-          if(file_exists($file['filename'] . $mkvm_ext)) {
-            unlink($file['filename'] . $mkvm_ext);
-          }
-        }        
-        stop($options, time());
-        return (array([], []));
-      }
-
-      if (file_exists($file['filename'] . $mkvm_ext) && file_exists($file['basename'])) {
-        if (empty($fileorig)) {
-          rename($file['basename'], $file['filename'] . ".orig." . $file['extension']);
-          $fileorig = pathinfo($file['filename'] . ".orig." . $file['extension']);
+      if (!$options['args']['test']) {
+        system("$cmdln 2>&1", $status);
+        if ($status == 255) {
+          // status(255) => CTRL-C
+          // Restore and Cleanup
+          if(!empty($fileorig) && !empty($file)) {
+            rename($fileorig['basename'], $file['basename']); 
+            if(file_exists($file['filename'] . $mkvm_ext)) {
+              unlink($file['filename'] . $mkvm_ext);
+            }
+          }        
+          stop($options, time());
+          return (array([], []));
         }
-        rename($file['filename'] . $mkvm_ext, $file['basename']);
-        $file     = pathinfo($file['basename']);
-        setXmlFormatAttribute($file, 'mkvmerged', true);
-        $options['args']['mkvmerged'] = true;
+        if (file_exists($file['filename'] . $mkvm_ext) && file_exists($file['basename'])) {
+          if (empty($fileorig)) {
+            rename($file['basename'], $file['filename'] . ".orig." . $file['extension']);
+            $fileorig = pathinfo($file['filename'] . ".orig." . $file['extension']);
+          }
+          rename($file['filename'] . $mkvm_ext, $file['basename']);
+          $file     = pathinfo($file['basename']);
+          setXmlFormatAttribute($file, 'mkvmerged', true);
+          $options['args']['mkvmerged'] = true;
+        }
       }
     }
   }
